pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "orquestra-agentes-financeiros"
version = "0.1.0"
description = "Sistema multi-agente para análise de risco de crédito usando LangGraph"
authors = ["Desenvolvedor <dev@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.0"
langgraph = "^0.0.40"
langchain-openai = "^0.0.8"
langchain-anthropic = "^0.1.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
requests = "^2.31.0"
aiohttp = "^3.9.0"
python-multipart = "^0.0.6"
pypdf = "^3.17.0"
python-docx = "^1.1.0"
pytesseract = "^0.3.10"
pillow = "^10.1.0"
faiss-cpu = "^1.7.4"
chromadb = "^0.4.18"
tiktoken = "^0.5.2"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
tavily-python = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
jupyter = "^1.0.0"
ipykernel = "^6.27.1"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"